Added mid-run splits
The splits have been tested as adequately as I was able to, but I don't know the run yet and I wrote them while running low on sleep, so there's a high chance that one or more of them might fail. I labelled them as being experimental because of this. The Silver Key split is missing because I don't know where in the level the key is, so I couldn't look for a split condition yet. The Flask of Lillandril split is similar, but I had a reasonable guess about what it should be able to split on so I went ahead and added that in.