Removed 3 cutscenes and updated pointers
Added the ship docking, ferry away, and ferry return cutscenes to the load removal, and after each cutscene it adds a set amount of time back to compensate for the varied cutscene length. The vmem base pointers for both Steam and patched GOG have been updated to be more reliable.